verified /combos is doing what it should
This because the original endpoint returned the set of garments that will be
used to generate a combo, and not existing combos themselves. To replicate
this behaviour, the endpoint [GET] /combos returns the list of combinations
that has already been produced; while a new endpoint /garmentsForCombos is
used to retrieve all garments so they can be used in combos.

Included also tests for these endpoints, as well as verified all endpoints
behave as expected from the originals.