[post] /garment is working!
added RGB-to-name library and approximation methods, as well as a method
to transform from RGB string to an RGB tuple.

This does not work for uploading images locally, though.

Regardless, the API should be deployable into AWS now. Unless we try the
lambda method to miniaturise the uploaded image file.